New rotational elements in use. Actual rotation (siderealTime) not properly working yet. A few other codefixes.

+++ src/core/modules/Comet.hpp
@@ -88,7 +88,7 @@ public:
 	//! for comets. They are used to calculate the apparent magnitude at
 	//! different distances from the Sun. They are not used in the same way
 	//! as the same parameters in MinorPlanet.
-	void setAbsoluteMagnitudeAndSlope(const double magnitude, const double slope);
+	void setAbsoluteMagnitudeAndSlope(const float magnitude, const float slope);
 
 	//! set value for semi-major axis in AU
 	void setSemiMajorAxis(const double value);
@@ -125,8 +125,8 @@ private:
 	//! @param xOffset for the dust tail, this may introduce a bend. Units are x per sqrt(z).
 	void computeParabola(const float parameter, const float topradius, const float zshift, QVector<Vec3d>& vertexArr, QVector<float>& texCoordArr, QVector<unsigned short>& indices, const float xOffset=0.0f);
 
-	double absoluteMagnitude;
-	double slopeParameter;
+	float absoluteMagnitude;
+	float slopeParameter;
 	double semiMajorAxis;
 	bool isCometFragment;
 	bool nameIsProvisionalDesignation;
@@ -145,7 +145,7 @@ private:
 	QVector<double> comaVertexArr;
 	QVector<float> comaTexCoordArr; //  --> 2014-08: could also be declared static, but it is filled by StelPainter...
 
-	// These are to avoid having index arrays for each comet when all are equal.
+	// These are static to avoid having index arrays for each comet when all are equal.
 	static bool createTailIndices;
 	static bool createTailTextureCoords;
